Modularize the API
