+++ pytvmaze/__init__.py
@@ -1,50 +1,1 @@
-#!/usr/bin/python
 # coding=utf-8
-
-from .tvmaze import (
-    AKA,
-    Cast,
-    CastCredit,
-    Character,
-    Crew,
-    CrewCredit,
-    Episode,
-    FollowedNetwork,
-    FollowedPerson,
-    FollowedShow,
-    FollowedWebChannel,
-    MarkedEpisode,
-    Person,
-    Season,
-    Show,
-    TVMaze,
-    Update,
-    Updates,
-    Network,
-    WebChannel,
-    episodes_by_date,
-    episode_by_id,
-    episode_by_number,
-    episode_list,
-    get_full_schedule,
-    get_people,
-    get_schedule,
-    get_show_crew,
-    get_show_list,
-    lookup_tvdb,
-    lookup_tvrage,
-    lookup_imdb,
-    people_search,
-    person_cast_credits,
-    person_crew_credits,
-    person_main_info,
-    season_by_id,
-    show_akas,
-    show_cast,
-    show_index,
-    show_main_info,
-    show_single_search,
-    show_search,
-    show_seasons,
-    show_updates,
-)
